Remove some lint warnings
git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/trunk@6431 1ca430b3-8271-4a6b-9b0e-3dcac98a8924

+++ nut/pro/uhttp/mtinit.c
@@ -94,7 +94,7 @@ int MediaTypeInitDefaults(void)
     int i;
 
     ISC_LIST_INIT(mediaTypeList);
-    for (i = 0; i < MT_DEFAULTS; i++) {
+    for (i = 0; i < (int)MT_DEFAULTS; i++) {
         ISC_LIST_APPEND(mediaTypeList, &mt_defaults[i], media_link);
     }
     return 0;
